Parse dates independent of system locale
Currently, the date parsing logic will pick up the system locale and can
thus fail to parse dates in those circumstances.  In this PR, the parser
is forced into the US locale.  Additionally, the tests which are subject
to the same bug are fixed.

With help from Jake McCrary @jakemcc